tests: update tests
Update tests to new structure, add new test cases. Specifically:

* Remove amount field: initially this was being used to sort the
  recipients and generated outputs, sorting is not specified by the
  protocol. Instead, remove the amount field and provide all possible
  output sets for the sending tests
* Add new test for outpoint sorting: add test case for lexicographically sorting
  outputs (vs sorting lexicographically by txid and then by integer
  vout)
* Add test cases for empty outputs and empty inputs

For the receiving tests, there was an additional check that the outputs
generated in the sending test match the given outputs in the receiving
test. This check is unnecessary and is removed. Now, the receiving test
takes the first "candidate set" from the expected outputs of the sending
test, so it is not guaranteed that the generated outputs from the
sending test will always match the given outputs for the receiving test.